+++ src/shared/Configuration.ts
@@ -3,8 +3,8 @@ import { name, publisher } from "../../package.json"
 
 // Base
 export const baseName = "Cline"
-export const baseVersion = "v3.30.3"
-export const baseVersionUrl = `https://github.com/cline/cline/blob/main/CHANGELOG.md#3303`
+export const baseVersion = "v3.32.5"
+export const baseVersionUrl = `https://github.com/cline/cline/blob/main/CHANGELOG.md#3325`
 
 // Product
 export const agentName = "Astro"