+++ CHANGELOG.md
@@ -1,11 +1,18 @@
 # Changelog
 
+## [1.2.0]
+
+- Task header design update for better usability
+
+Base Features Update:
+- Claude Sonnet 4.5: Support for Sonnet 4.5 in multiple API providers. [v3.32.5](https://github.com/cline/cline/blob/main/CHANGELOG.md#3325)
+
 ## [1.1.0]
 
-Base Cline Features Update:
+Base Features Update:
 - Focus Chain: Automatically creates and maintains todo lists in tasks, breaking down tasks into steps with real-time progress tracking
 - Deep Planning: New "/Deep Planning" slash command for codebase exploration and implementation planning, that integrates with Focus Chain for automatic progress tracking
-- New API providers, models, enhancements and fixes: v3.30.3
+- New API providers, models, enhancements and fixes. [v3.30.3](https://github.com/cline/cline/blob/main/CHANGELOG.md#3303)
 
 ## [1.0.1]